Rename "setupErdjsSnippets" to "setupJavaScriptSnippets".
