+++ package.json
@@ -110,8 +110,8 @@
 				"category": "multiversx"
 			},
 			{
-				"command": "multiversx.setupErdjsSnippets",
-				"title": "Setup erdjs-snippets",
+				"command": "multiversx.setupJavaScriptSnippets",
+				"title": "Setup JavaScript snippets",
 				"category": "multiversx"
 			}
 		],
@@ -119,7 +119,7 @@
 			"explorer/context": [
 				{
 					"command": "multiversx.buildContract",
-					"when": "explorerResourceIsFolder && resourceFilename != output && resourceFilename != debug && resourceFilename != target && resourceFilename != erdjs-snippets && resourceFilename != interaction",
+					"when": "explorerResourceIsFolder && resourceFilename != output && resourceFilename != debug && resourceFilename != target && resourceFilename != js-snippets && resourceFilename != interaction",
 					"group": "multiversx"
 				},
 				{
@@ -129,7 +129,7 @@
 				},
 				{
 					"command": "multiversx.cleanContract",
-					"when": "explorerResourceIsFolder && resourceFilename != output && resourceFilename != debug && resourceFilename != target && resourceFilename != erdjs-snippets && resourceFilename != interaction",
+					"when": "explorerResourceIsFolder && resourceFilename != output && resourceFilename != debug && resourceFilename != target && resourceFilename != js-snippets && resourceFilename != interaction",
 					"group": "multiversx"
 				},
 				{
@@ -139,7 +139,7 @@
 				},
 				{
 					"command": "multiversx.runContractSnippet",
-					"when": "explorerResourceIsFolder && resourceFilename != output && resourceFilename != debug && resourceFilename != target && resourceFilename != erdjs-snippets",
+					"when": "explorerResourceIsFolder && resourceFilename != output && resourceFilename != debug && resourceFilename != target && resourceFilename != js-snippets",
 					"group": "multiversx"
 				},
 				{
@@ -154,7 +154,7 @@
 				},
 				{
 					"command": "multiversx.runScenarios",
-					"when": "explorerResourceIsFolder && resourceFilename != output && resourceFilename != debug && resourceFilename != target && resourceFilename != erdjs-snippets && resourceFilename != interaction",
+					"when": "explorerResourceIsFolder && resourceFilename != output && resourceFilename != debug && resourceFilename != target && resourceFilename != js-snippets && resourceFilename != interaction",
 					"group": "multiversx"
 				},
 				{
