+++ src/javaScriptSnippets.ts
